+++ CotEditor/Sources/AppearancePaneController.swift
@@ -42,9 +42,16 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
     
     @IBOutlet private weak var fontField: AntialiasingTextField?
     @IBOutlet private weak var lineHeightField: NSTextField?
+    
     @IBOutlet private weak var barCursorButton: NSButton?
     @IBOutlet private weak var thickBarCursorButton: NSButton?
     @IBOutlet private weak var blockCursorButton: NSButton?
+    
+    @IBOutlet private weak var appearanceLabelField: NSTextField?
+    @IBOutlet private weak var defaultAppearanceButton: NSButton?
+    @IBOutlet private weak var lightAppearanceButton: NSButton?
+    @IBOutlet private weak var darkAppearanceButton: NSButton?
+    
     @IBOutlet private weak var themeTableView: NSTableView?
     @IBOutlet private var themeTableMenu: NSMenu?
     
@@ -67,6 +74,14 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
         // set initial value as field's placeholder
         self.lineHeightField?.bindNullPlaceholderToUserDefaults(.value)
         
+        // remove appearance controls on macOS 10.13 or earlier
+        if NSAppKitVersion.current <= .macOS10_13 {
+            self.appearanceLabelField?.removeFromSuperview()
+            self.defaultAppearanceButton?.removeFromSuperview()
+            self.lightAppearanceButton?.removeFromSuperview()
+            self.darkAppearanceButton?.removeFromSuperview()
+        }
+        
         // observe theme list change
         NotificationCenter.default.addObserver(self, selector: #selector(setupThemeList), name: didUpdateSettingListNotification, object: ThemeManager.shared)
         NotificationCenter.default.addObserver(self, selector: #selector(themeDidUpdate), name: didUpdateSettingNotification, object: ThemeManager.shared)
@@ -88,6 +103,16 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
             self.thickBarCursorButton?.state = .on
         case .block:
             self.blockCursorButton?.state = .on
+        }
+        
+        // select one of appearance radio buttons
+        switch UserDefaults.standard[.documentAppearance] {
+        case .default:
+            self.defaultAppearanceButton?.state = .on
+        case .light:
+            self.lightAppearanceButton?.state = .on
+        case .dark:
+            self.darkAppearanceButton?.state = .on
         }
         
         let themeName = ThemeManager.shared.userDefaultSettingName(forDark: self.view.effectiveAppearance.isDark)
@@ -377,6 +402,13 @@ final class AppearancePaneController: NSViewController, NSMenuItemValidation, NS
     @IBAction func updateCursorTypeSetting(_ sender: NSButton) {
         
         UserDefaults.standard[.cursorType] = CursorType(rawValue: sender.tag)!
+    }
+    
+    
+    /// A radio button of documentAppearance was clicked
+    @IBAction func updateAppearanceSetting(_ sender: NSButton) {
+        
+        UserDefaults.standard[.documentAppearance] = AppearanceMode(rawValue: sender.tag)!
     }