+++ CotEditor/Sources/DefaultSettings.swift
@@ -9,7 +9,7 @@
 //  ---------------------------------------------------------------------------
 //
 //  © 2004-2007 nakamuxu
-//  © 2014-2018 1024jp
+//  © 2014-2019 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -80,6 +80,7 @@ struct DefaultSettings {
         .showInvisibleFullwidthSpace: false,
         .invisibleFullwidthSpace: 0,
         .showOtherInvisibleChars: false,
+        .documentAppearance: AppearanceMode.default.rawValue,
 //        .theme: "Dendrobates",  // -> The default theme is set dynamically by taking the current appearance into account.
         
         .smartInsertAndDelete: false,